refactor: interfaceとデフォルト実装を同じファイルに
